吸魂フィールドをとりあえず実装
吸魂フィールドをとりあえず実装
後で魂バート弾時に大きくなる処理を入れる予定

+++ Assets/Scripts/EnemySetting.cs
@@ -12,8 +12,8 @@ using UnityEngine;
 public class EnemyData
 {
     public string Id;
-    public int Hp;
-    public int Attack;
+    public float Hp;
+    public float Attack;
     public int Money;
 }
